agent/requirements.txt
streamlit
langchain
langchain-community
langchain-qdrant
langchain-google-genai
langgraph
fastembed
qdrant-client
python-dotenv